v1/14-supermarket-shoppinglist/requirements.txt
streamlit
google-generativeai
google-cloud-aiplatform
python-dotenv
langchain_community
langchain
google-cloud-discoveryengine
streamlit-carousel